# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OEE System API Documentation",
    "version": "1.0.0",
    "description": "Welcome to the OEE (Overall Equipment Effectiveness) System API documentation..."
  },
  "servers": [
    {
      "url": "https://localhost/api/v1"
    }
  ],
  "paths": {
    "/machines": {
      "get": {
        "summary": "Get all machines",
        "tags": [
          "Machines"
        ],
        "description": "Retrieve a list of all machines.",
        "responses": {
          "200": {
            "description": "A list of machines.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new machine",
        "tags": [
          "Machines"
        ],
        "description": "Create a new machine.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Machine created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "machine": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/microstop-aggregation/machine": {
      "get": {
        "summary": "Get aggregated microstop data, optionally filtered by machine ID and date range",
        "tags": [
          "Microstop Aggregation"
        ],
        "description": "Retrieves an aggregation of microstop data, optionally filtered by a specific machine ID and date range.",
        "parameters": [
          {
            "in": "query",
            "name": "machine_id",
            "required": false,
            "description": "The unique identifier of the machine to filter the microstops by.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "description": "Start date for filtering microstops, in ISO 8601 format.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "description": "End date for filtering microstops, in ISO 8601 format.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved aggregated microstop data."
          },
          "404": {
            "description": "No data found for the given filters."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/microstop-aggregation/process-order": {
      "get": {
        "summary": "Get aggregated microstops by process order",
        "description": "Retrieve microstops aggregated by reason code for a given process order number, or filter by start and end date.",
        "tags": [
          "Microstop Aggregation"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "processOrderNumber",
            "schema": {
              "type": "string"
            },
            "description": "The process order number to filter by."
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start date to filter microstops (format: YYYY-MM-DDTHH:mm:ssZ)."
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End date to filter microstops (format: YYYY-MM-DDTHH:mm:ssZ)."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of aggregated microstops.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Microstop not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/microstops": {
      "get": {
        "summary": "Get all microstops",
        "tags": [
          "Microstops"
        ],
        "description": "Retrieve a list of all microstops.",
        "responses": {
          "200": {
            "description": "A list of microstops.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new microstop",
        "tags": [
          "Microstops"
        ],
        "description": "Create a new microstop.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "Start": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "End": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Microstop added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "microstop": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/oeeconfig": {
      "get": {
        "summary": "Get the entire OEE configuration",
        "tags": [
          "OEE Configuration"
        ],
        "description": "Retrieve the entire OEE configuration as a JSON object.",
        "responses": {
          "200": {
            "description": "The entire OEE configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new OEE configuration",
        "tags": [
          "OEE Configuration"
        ],
        "description": "Add a new key-value pair to the OEE configuration.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New OEE configuration added successfully."
          }
        }
      }
    },
...
    "/planneddowntime": {
      "get": {
        "summary": "Get all planned downtimes",
        "tags": [
          "Planned Downtime"
        ],
        "description": "Retrieve a list of all planned downtimes.",
        "responses": {
          "200": {
            "description": "A list of planned downtimes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new planned downtime",
        "tags": [
          "Planned Downtime"
        ],
        "description": "Create a new planned downtime.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Planned downtime added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/processorders": {
      "get": {
        "summary": "Get all process orders",
        "tags": [
          "Process Orders"
        ],
        "description": "Retrieve a list of all process orders.",
        "responses": {
          "200": {
            "description": "A list of process orders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new process order",
        "tags": [
          "Process Orders"
        ],
        "description": "Create a new process order.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Process order added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ratings": {
      "get": {
        "summary": "Get all ratings",
        "tags": [
          "Ratings"
        ],
        "description": "Retrieve a list of all ratings.",
        "responses": {
          "200": {
            "description": "A list of ratings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new rating",
        "tags": [
          "Ratings"
        ],
        "description": "Create a new rating.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rating created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/shiftmodels": {
      "get": {
        "summary": "Get all shift models",
        "tags": [
          "Shift Models"
        ],
        "description": "Retrieve a list of all shift models.",
        "responses": {
          "200": {
            "description": "A list of shift models.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new shift model",
        "tags": [
          "Shift Models"
        ],
        "description": "Create a new shift model.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shift model added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "shiftModel": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/structure": {
      "get": {
        "summary": "Fetch the structure configuration",
        "tags": [
          "Structure"
        ],
        "description": "Reads the configuration from the structure.json file and sends it as a response.",
        "responses": {
          "200": {
            "description": "The structure configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Update the structure configuration",
        "tags": [
          "Structure"
        ],
        "description": "Writes the new configuration to the structure.json file.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structure configuration saved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tact": {
      "get": {
        "summary": "Get all tact entries",
        "tags": [
          "Tact"
        ],
        "description": "Retrieve a list of all tact entries.",
        "responses": {
          "200": {
            "description": "A list of tact entries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new tact entry",
        "tags": [
          "Tact"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "machine": {
                    "type": "string"
                  },
                  "material": {
                    "type": "string"
                  },
                  "sollMax": {
                    "type": "number"
                  },
                  "sollMin": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tact entry created successfully."
          }
        }
      }
    },
...
    "/topics": {
      "get": {
        "summary": "Get dynamic topics",
        "tags": [
          "Topics"
        ],
        "description": "Generate topics based on the Plant, Area, or Line.",
        "parameters": [
          {
            "in": "query",
            "name": "plant",
            "schema": {
              "type": "string"
            },
            "description": "Filter by Plant"
          },
          {
            "in": "query",
            "name": "area",
            "schema": {
              "type": "string"
            },
            "description": "Filter by Area"
          },
          {
            "in": "query",
            "name": "line",
            "schema": {
              "type": "string"
            },
            "description": "Filter by Line"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of generated topics.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/unplanneddowntime": {
      "get": {
        "summary": "Get all unplanned downtimes",
        "tags": [
          "Unplanned Downtime"
        ],
        "description": "Retrieve a list of all unplanned downtimes.",
        "responses": {
          "200": {
            "description": "A list of unplanned downtimes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new unplanned downtime",
        "tags": [
          "Unplanned Downtime"
        ],
        "description": "Create a new unplanned downtime.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Unplanned downtime added successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Get all users",
        "tags": [
          "Users"
        ],
        "description": "Retrieve a list of all users.",
        "responses": {
          "200": {
            "description": "A list of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": [
          "Users"
        ],
        "description": "Create a new user and save it to the list.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}